Refactored the KeyBindingBus as an Extension to the EventBus, instead of his own Bus
