+++ src/main/java/com/ddoerr/scriptit/screens/BindingScreen.java
@@ -33,7 +33,7 @@ public class BindingScreen extends Screen {
         children.add(keyBindingsListWidget);
 
         ButtonWidget addButton = new ButtonWidget(window.getScaledWidth() - 220, window.getScaledHeight() - 35, 200, 20, I18n.translate("scriptit:bindings.add"), (button) -> {
-            Trigger trigger = new BusTrigger(KeyBindingBus.class.getSimpleName(), InputUtil.UNKNOWN_KEYCODE.getName());
+            Trigger trigger = new BusTrigger(KeyBindingBusExtension.class.getSimpleName(), InputUtil.UNKNOWN_KEYCODE.getName());
             ScriptContainer scriptContainer = new ScriptContainer(trigger);
             Resolver.getInstance().resolve(Scripts.class).add(Scripts.KEYBIND_CATEGORY, scriptContainer);
